fixes c api build on windows

+++ setup.py
@@ -9,67 +9,79 @@ from setuptools import Extension, setup
 from setuptools.command.build_ext import build_ext
 from setuptools.command.install_lib import install_lib
 
-
 ext_clean = os.environ.get('ASSESS_EXT_CLEAN')
 
 assess_ext_path = os.path.join('dongtai_agent_python', 'assess_ext')
 c_sources = glob(os.path.join(assess_ext_path, '*.c')) + glob(os.path.join(assess_ext_path, "patch/*.c"))
 
-if platform.startswith('darwin'):
-    link_args = ['-rpath', '@loader_path']
-    platform_args = []
-else:
-    platform_args = ['-Wno-cast-function-type']
-    link_args = []
-
-strict_build_args = ['-Werror']
-
-
-extensions = [
-    Extension(
-        'dongtai_agent_python.assess_ext.c_api',
-        c_sources,
-        libraries=["funchook"],
-        include_dirs=[
-            os.path.join(assess_ext_path, 'include'),
-            os.path.join(assess_ext_path, 'funchook/include'),
-        ],
-        library_dirs=[os.path.join(assess_ext_path, 'funchook/build')],
-        runtime_library_dirs=['$ORIGIN'],
-        extra_compile_args=[
-            '-Wall',
-            '-Wextra',
-            '-Wno-unused-parameter',
-            '-Wmissing-field-initializers',
-        ]
-        + strict_build_args
-        + platform_args,
-        extra_link_args=link_args,
-    )
-]
-
-
 tempdir = None
 funchook_temp = None
+
+include_dirs = [
+    os.path.join(assess_ext_path, 'include'),
+    os.path.join(assess_ext_path, 'funchook', 'include'),
+]
+
+runtime_library_dirs = []
+link_args = []
+extra_compile_args = []
+strict_build_args = []
+platform_args = []
+
+if platform.startswith('linux'):
+    platform_args = ['-Wno-cast-function-type']
+elif platform.startswith('darwin'):
+    link_args = ['-rpath', '@loader_path']
+
+if platform == 'win32':
+    libraries = ['funchook', 'distorm', 'psapi']
+    library_dirs = [
+        os.path.join(assess_ext_path, 'funchook', 'build', 'Release'),
+    ]
+else:
+    libraries = ['funchook']
+    library_dirs = [
+        os.path.join(assess_ext_path, 'funchook', 'build'),
+    ]
+    runtime_library_dirs = ['$ORIGIN']
+    extra_compile_args = [
+        '-Wall',
+        '-Wextra',
+        '-Wno-unused-parameter',
+        '-Wmissing-field-initializers',
+    ]
+    strict_build_args = ['-Werror']
 
 
 class BuildExt(build_ext):
     def run(self):
-        build_script = '''
-        cd %s
-        mkdir -p build
-        cd build
-        %s
+        cmake_cmd = '''
         cmake -DCMAKE_BUILD_TYPE=Release ..
         make
-        ''' % (
-            os.path.join(assess_ext_path, 'funchook'),
-            'make clean' if ext_clean else '',
-        )
+        '''
+        if platform == 'win32':
+            cmake_cmd = '''
+            cmake.exe  -DCMAKE_INSTALL_PREFIX=%s ..
+            cmake --build . --config Release
+            ''' % os.path.join(assess_ext_path, 'funchook')
+
+        build_dir = os.path.join(assess_ext_path, 'funchook/build')
+        if ext_clean and os.path.exists(build_dir):
+            rmtree(build_dir)
+        if not os.path.exists(build_dir):
+            os.mkdir(build_dir)
+
+        build_script = '''
+        cd %s
+        %s
+        ''' % (build_dir, cmake_cmd)
         if os.system(build_script) != 0:
             raise DistutilsExecError("Failed to build DongTai-agent-python C extension")
 
         build_ext.run(self)
+
+        if platform == 'win32':
+            return
 
         global tempdir
         global funchook_temp
@@ -87,15 +99,36 @@ class InstallLib(install_lib):
     def run(self):
         install_lib.run(self)
 
+        if platform == 'win32':
+            return
+
+        global tempdir
+        global funchook_temp
+
         if funchook_temp is not None:
             dest_dir = os.path.join(self.install_dir, assess_ext_path)
             self.copy_file(funchook_temp, dest_dir)
             rmtree(tempdir)
 
 
+extensions = [
+    Extension(
+        'dongtai_agent_python.assess_ext.c_api',
+        c_sources,
+        libraries=libraries,
+        include_dirs=include_dirs,
+        library_dirs=library_dirs,
+        runtime_library_dirs=runtime_library_dirs,
+        extra_compile_args=extra_compile_args + strict_build_args + platform_args,
+        extra_link_args=link_args,
+    )
+]
+
+
 setup(
     name='dongtai_agent_python',
-    scripts=['scripts/dongtai-cli'],
+    scripts=['scripts' + os.sep + 'dongtai-cli'],
+    include_package_data=True,
     cmdclass=dict(build_ext=BuildExt, install_lib=InstallLib),
     ext_modules=extensions,
     entry_points={
